backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
transformers==4.35.2
torch==2.1.1
easyocr==1.7.0
pillow==10.1.0
speechrecognition==3.10.0
pydub==0.25.1
aiohttp==3.9.1
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0